Extract Gdi class from OpenGL.cs into dedicated Gdi.cs
Pure file-split refactor: moves the GDI32 P/Invoke bindings out of the
52k-line OpenGL.cs into CADability/Gdi.cs so a future deletion of the
legacy OpenGL layer does not break Text.cs or PaintToOpenGL.cs.
No logic changes; namespace and all callers are unaffected.
